requirements.txt
pillow==10.3.0
pypng==0.20220715.0
qrcode==7.4.2
typing-extensions==4.10.0
